api/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
click==8.0.1
colorama==0.4.6
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.14
Werkzeug==2.0.1
WTForms==2.3.3
boto3 
matplotlib
openai==0.19.0
pycryptodome